Visualizing Clusters using PCA
What we've done is we've taken those columns of 0/1 indicator variables, and we've transformed them into a 2-D dataset. We took one column and arbitrarily called it x and then called the other y. We threw each point into a scatterplot. We color coded each point based on it's cluster so it's easier to see them.

Ultimately, K = 3 did a good job at clustering than K = 4. We can see it visually by above plots. So optimum value for K is 3. Hence, the Elbow method of choosing 𝐾 seems to have produced the optimal result visually.